refactor(token-controls): move parseCssLength into the shared token-control library
