+++ src/token-controls/__tests__/parse-css-length.test.js
@@ -1,6 +1,6 @@
 /* eslint-env jest */
 
-import { parseCssLength } from '../parse-css-length';
+import { parseCssLength } from '../helpers/parse-css-length';
 
 describe('parseCssLength', () => {
 	it.each([
